add checking and timeout

+++ programming/devel-setup.sh
@@ -19,6 +19,7 @@ add_path() {
 temp=$(mktemp -d)
 
 echo ">>>> C/C++"
+
 os=$(uname -s)
 case $os in
     Darwin)
@@ -26,56 +27,64 @@ case $os in
 		brew install cmake pkg-config || exit 1
 		;;
     Linux)
-		which apt > /dev/null && sudo apt install -y clang cmake pkg-config
-		which dnf > /dev/null && sudo dnf install -y clang cmake pkg-config
+        cc_list="gcc g++ clang cmake pkg-config build-essential"
+		which apt > /dev/null && sudo apt install -y $cc_list
+		which dnf > /dev/null && sudo dnf install -y $cc_list
 		;;
     *)
 		echo "$os not supported! (skipped)"
 esac
 
-cd $temp
-while [ ! -d fmt ]
-do
-    git clone https://github.com/fmtlib/fmt.git
-done
+if [ ! pkg-config --exists fmt ]; then
+    cd $temp
+    while [ ! -d fmt ]
+    do
+        git clone https://github.com/fmtlib/fmt.git
+    done
 
-mkdir -vp fmt/build
-cd fmt/build
-cmake ..
-make -j4 && sudo make install
-
-if [ $? -ne 0 ]; then
-    echo "fail to install fmt!"
-    exit 1
+    mkdir -vp fmt/build && cd fmt/build
+    cmake ..
+    make -j4 && sudo make install
 fi
 
 echo ">>>> C#"
-curl -sfSL https://dot.net/v1/dotnet-install.sh | bash
-grep DOTNET_ROOT $profile || echo 'export DOTNET_ROOT=$HOME/.dotnet' >> $profile
-add_path '$HOME/.dotnet:$HOME/.dotnet/tools'
+
+if ! dotnet --info | head -10; then
+    curl -sfSL https://dot.net/v1/dotnet-install.sh | bash
+    grep DOTNET_ROOT $profile || echo 'export DOTNET_ROOT=$HOME/.dotnet' >> $profile
+    add_path '$HOME/.dotnet:$HOME/.dotnet/tools'
+fi
 
 echo ">>>> Dart"
-cd ~
-while [ ! -d flutter/bin ]
+
+flutter_path=$HOME/flutter
+while [ ! -x $flutter_path/bin/dart ]
 do
-    rm -rf flutter
-    git clone https://github.com/flutter/flutter.git -b stable
+    rm -rf $flutter_path
+    git clone https://github.com/flutter/flutter.git $flutter_path -b stable
 done
 
-add_path `pwd`/flutter/bin
+add_path $flutter_path/bin
 
 echo ">>>> Go"
+
 mkdir -p $HOME/bin
 add_path $HOME/bin
 
 while [ ! -x $HOME/bin/gvm ]
 do
-    curl -L -O https://github.com/devnw/gvm/releases/download/latest/gvm && install -v -m 755 gvm $HOME/bin/
+    cd $temp && rm -rf gvm
+    curl -L -O https://github.com/devnw/gvm/releases/download/latest/gvm && \
+        install -v -m 755 gvm $HOME/bin/
 done
 
-#$HOME/bin/gvm 1.19.3
+# FIX gvm
+if [ $(uname -m) == x86_64 ]; then
+    $HOME/bin/gvm next
+fi
 
 echo ">>>> Java & Kotlin"
+
 while [ ! -e ~/.sdkman/bin/sdkman-init.sh ]
 do
     rm -rf ~/.sdkman
@@ -86,16 +95,22 @@ source ~/.sdkman/bin/sdkman-init.sh
 # FIXME
 for pkg in java kotlin maven gradle
 do
-    while true
+    for ((i=0;i<10;i++))
 	do
         sdk install $pkg && break
     done
 done
 
-echo ">>>> JavaScript: Deno"
-curl -fsSL https://deno.land/install.sh | sh
+echo ">>>> JavaScript: deno"
 
-echo ">>>> JavaScript: NVM"
+for ((i=0;i<5;i++))
+do
+    curl -fsSL https://deno.land/install.sh | sh
+    which deno && break
+done
+
+echo ">>>> JavaScript: node"
+
 while [ ! ~/.nvm/nvm.sh ]
 do
     cd $temp
@@ -108,25 +123,34 @@ done
 
 export NVM_DIR="$HOME/.nvm"
 [ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"  # This loads nvm
-[ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion"  # This loads nvm bash_completion
+# [ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion"  # This loads nvm bash_completion
 
 nvm install --lts
 
 echo ">>>> Rust"
-which rustc || curl --proto '=https' --tlsv1.3 https://sh.rustup.rs -sSf | sh
+
+if ! rustc --version; then
+    for ((i=0;i<10;i++))
+    do
+        curl --proto '=https' --tlsv1.3 https://sh.rustup.rs -sSf | sh
+        [ $? -eq 0 ] && break
+    done
+fi
 
 echo ">>>> Swift"
 
-case $(uname -s) in
-Darwin)
-    brew install docker
-    ;;
-Linux)
-    curl -fsSL https://get.docker.com | bash -s docker --mirror Aliyun
-    sudo usermod -aG docker $USER
-    newgrp docker
-    ;;
-esac
+if ! docker --version; then
+    case $(uname -s) in
+    Darwin)
+        brew install docker
+        ;;
+    Linux)
+        curl -fsSL https://get.docker.com | bash -s docker --mirror Aliyun
+        sudo usermod -aG docker $USER
+        newgrp docker
+        ;;
+    esac
+fi
 
 for ((i=0;i<10;i++))
 do
